requirements.txt
# Core dependencies for Contextual Retrieval System

# Anthropic Claude API
anthropic>=0.39.0

# Vector database
qdrant-client>=1.7.0

# Embeddings
sentence-transformers>=2.2.2

# BM25 for lexical search
rank-bm25>=0.2.2

# Text processing
tiktoken>=0.5.2

# Utilities
python-dotenv>=1.0.0
numpy>=1.24.0
